refactor: google module api connections
