+++ .env.example
@@ -22,7 +22,5 @@ GOOGLE_API_SCOPE="https://www.googleapis.com/auth/drive"
 GOOGLE_DOC_BASE_URL="https://docs.google.com/document/d/"
 GOOGLE_DRIVE_FOLDER_ID=
 GOOGLE_DOC_TEMPLATE_ID=
-GOOGLE_DRIVE_API_VERSION=v3
-GOOGLE_DOC_API_VERSION=v1
 GOOGLE_DOC_TEMPLATE_MAX_PROJECTS=5
 GOOGLE_DOC_TEMPLATE_FIELDS='{"professionalName":"{{name}}","professionalTechnologies":"{{technologies}}","professionalHeadline":"{{headline}}","professionalAboutMe":"{{about_me}}","professionalEnglishLevel":"{{english_level}}","projectTitle":"{{project_title_{{index}}}}","projectDescription":"{{project_description_{{index}}}}","projectTechnologiesTitle":"{{project_technologies_title_{{index}}}}","projectTechnologiesTitleText":"Technologies: ","projectTechnologies":"{{project_technologies_{{index}}}}","projectNewLine": "\n\n","indexTextForReplace":"{{index}}"}'
