+++ src/dtos/res/build-resume-res.dto.ts
@@ -1,6 +1,8 @@
 import { ApiProperty } from '@nestjs/swagger';
+import { Expose } from 'class-transformer';
 
 export class BuildResumeResDto {
   @ApiProperty({ type: String })
+  @Expose()
   public resumeUrl?: string;
 }
