+++ src/resume-builder/__test__/resume-builder.service.spec.ts
@@ -18,8 +18,6 @@ describe('ResumeBuilderService', () => {
             googleDocBaseUrl: 'googleDocBaseUrl',
             googleDriveFolderId: 'googleDriveFolderId',
             googleDocTemplateId: 'googleDocTemplateId',
-            googleDriveApiVersion: 'googleDriveApiVersion',
-            googleDocApiVersion: 'googleDocApiVersion',
             googleDocTemplateMaxProjects: 'googleDocTemplateMaxProjects',
             googleDocTemplateFields: 'googleDocTemplateFields',
           },
